+++ src/index.ts
@@ -1,18 +1,37 @@
-import Button from "components/Button";
-import Input, { InputType } from "components/Input";
-import Loader, { LoaderType } from "components/Loader";
-import Modal, { ModalType } from "components/Modal";
-import MultiStepForm from "components/MultiStepForm";
-import Switch from "components/Switch";
+export * as Footer from "components/Footer";
+export type {
+  ContactInfoProps,
+  CopyrightProps,
+  FooterProps,
+  LinkProps,
+} from "components/Footer";
 
-export default {
-  Button,
-  Input,
-  InputType,
-  Loader,
-  LoaderType,
-  Modal,
-  ModalType,
-  MultiStepForm,
-  Switch,
-};
+export * as Input from "components/Input";
+export { InputType, type InputProps } from "components/Input";
+
+export * as Loader from "components/Loader";
+export { LoaderType, type LoaderProps } from "components/Loader";
+
+export * as Modal from "components/Modal";
+export { ModalType, type ModalProps } from "components/Modal";
+
+export * as MultiStepForm from "components/MultiStepForm";
+export type { MultiStepProps } from "components/MultiStepForm";
+
+export * as SocialMediaLink from "components/SocialMediaLink";
+export {
+  Platform as SocialMediaPlatform,
+  type SocialMediaArgs,
+} from "components/SocialMediaLink";
+
+export * as SocialMediaList from "components/SocialMediaList";
+export type {
+  SocialMediaListDisplayProps,
+  SocialMediaListProps,
+} from "components/SocialMediaList";
+
+export * as Switch from "components/Switch";
+export type { ToggleProps } from "components/Switch";
+
+export * as QRCode from "components/QRCode";
+export type { QRCodeDisplayOptions, QRCodeProps } from "components/QRCode";